Support math in Chrome.
In Chrome, we use exactly the same code to retrieve MathML as we do in Firefox, so move the math code from NVDAObjects.IAccessible.mozilla into NVDAObjects.IAccessible.ia2Web (which is used by both Firefox and Chrome). We also catch COMError because this isn't yet in Chrome stable and because this can fail in Firefox too if the ISimpleDOM COM proxy isn't registered.
