+++ source/NVDAObjects/IAccessible/mozilla.py
@@ -3,14 +3,13 @@
 #A part of NonVisual Desktop Access (NVDA)
 #This file is covered by the GNU General Public License.
 #See the file COPYING for more details.
-#Copyright (C) 2006-2015 NV Access Limited, Peter Vágner
+#Copyright (C) 2006-2017 NV Access Limited, Peter Vágner
 
 from collections import namedtuple
-from ctypes import c_short
 import IAccessibleHandler
 import oleacc
 import winUser
-from comtypes import IServiceProvider, COMError, BSTR
+from comtypes import IServiceProvider, COMError
 import eventHandler
 import controlTypes
 from . import IAccessible, Dialog, WindowRoot
@@ -156,27 +155,6 @@ class TextLeaf(Mozilla):
 	role = controlTypes.ROLE_STATICTEXT
 	beTransparentToMouse = True
 
-class Math(Mozilla):
-
-	def _get_mathMl(self):
-		from comtypes.gen.ISimpleDOM import ISimpleDOMNode
-		node = self.IAccessibleObject.QueryInterface(ISimpleDOMNode)
-		# Try the data-mathml attribute.
-		attr = node.attributesForNames(1, (BSTR * 1)("data-mathml"), (c_short * 1)(0,))
-		if attr:
-			import mathPres
-			if not mathPres.getLanguageFromMath(attr) and self.language:
-				attr = mathPres.insertLanguageIntoMath(attr, self.language)
-			return attr
-		if self.IA2Attributes.get("tag") != "math":
-			# This isn't MathML.
-			raise LookupError
-		if self.language:
-			attrs = ' xml:lang="%s"' % self.language
-		else:
-			attrs = ""
-		return "<math%s>%s</math>" % (attrs, node.innerHTML)
-
 def findExtraOverlayClasses(obj, clsList):
 	"""Determine the most appropriate class if this is a Mozilla object.
 	This works similarly to L{NVDAObjects.NVDAObject.findOverlayClasses} except that it never calls any other findOverlayClasses method.
@@ -238,7 +216,6 @@ _IAccessibleRolesToOverlayClasses = {
 	IAccessibleHandler.IA2_ROLE_EMBEDDED_OBJECT: EmbeddedObject,
 	"embed": EmbeddedObject,
 	"object": EmbeddedObject,
-	oleacc.ROLE_SYSTEM_EQUATION: Math,
 }
 
 #: Roles that mightn't set the focused state when they are focused.
